Update Program.cs
changed the Btn function to add a button name. Also changed First_Click & Second_Click.

First_Click - the button will be written as occupied in the file.
In Second_Click -  it will rewrite the file, its status will be rewritten to unoccupied. In the old check parking space, it counts the data in the file, even if it's unoccupied. It's fixed now.